requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
langchain==0.1.0
langchain-groq==0.0.1
faiss-cpu==1.7.4
nomic==2.0.3
python-dotenv==1.0.0
